+++ packages/plugin-vue-jsx/src/index.ts
@@ -13,7 +13,10 @@ export * from './types'
 const ssrRegisterHelperId = '/__vue-jsx-ssr-register-helper'
 const ssrRegisterHelperCode =
   `import { useSSRContext } from "vue"\n` +
-  `export ${ssrRegisterHelper.toString()}`
+  // the const here is just to work around the Bun bug where
+  // Function.toString() isn't working as intended
+  // https://github.com/oven-sh/bun/issues/9543
+  `export const ssrRegisterHelper = ${ssrRegisterHelper.toString()}`
 
 /**
  * This function is serialized with toString() and evaluated as a virtual
@@ -41,6 +44,7 @@ function vueJsxPlugin(options: Options = {}): Plugin {
     exclude,
     babelPlugins = [],
     defineComponentName = ['defineComponent'],
+    tsPluginOptions = {},
     ...babelPluginOptions
   } = options
   const filter = createFilter(include || /\.[jt]sx$/, exclude)
@@ -97,7 +101,7 @@ function vueJsxPlugin(options: Options = {}): Plugin {
               (r) => r.default,
             ),
             // @ts-ignore
-            { isTSX: true, allowExtensions: true },
+            { ...tsPluginOptions, isTSX: true, allowExtensions: true },
           ])
         }