+++ packages/plugin-vue/src/index.ts
@@ -1,6 +1,6 @@
 import fs from 'node:fs'
 import type { Plugin, ViteDevServer } from 'vite'
-import { createFilter } from 'vite'
+import { createFilter, normalizePath } from 'vite'
 /* eslint-disable import/no-duplicates */
 import type {
   SFCBlock,
@@ -19,7 +19,7 @@ import {
   getSrcDescriptor,
   getTempSrcDescriptor,
 } from './utils/descriptorCache'
-import { clearScriptCache, getResolvedScript, typeDepToSFCMap } from './script'
+import { clearScriptCache, resolveScript, typeDepToSFCMap } from './script'
 import { transformMain } from './main'
 import { handleHotUpdate, handleTypeDepChange } from './handleHotUpdate'
 import { transformTemplateAsModule } from './template'
@@ -217,6 +217,12 @@ export default function vuePlugin(rawOptions: Options = {}): Plugin<Api> {
     },
 
     handleHotUpdate(ctx) {
+      ctx.server.ws.send({
+        type: 'custom',
+        event: 'file-changed',
+        data: { file: normalizePath(ctx.file) },
+      })
+
       if (options.value.compiler.invalidateTypeCache) {
         options.value.compiler.invalidateTypeCache(ctx.file)
       }
@@ -317,7 +323,12 @@ export default function vuePlugin(rawOptions: Options = {}): Plugin<Api> {
         let block: SFCBlock | null | undefined
         if (query.type === 'script') {
           // handle <script> + <script setup> merge via compileScript()
-          block = getResolvedScript(descriptor, ssr)
+          block = resolveScript(
+            descriptor,
+            options.value,
+            ssr,
+            customElementFilter.value(filename),
+          )
         } else if (query.type === 'template') {
           block = descriptor.template!
         } else if (query.type === 'style') {