+++ packages/plugin-vue/src/main.ts
@@ -50,7 +50,7 @@ export async function transformMain(
       // for vue files, create descriptor from fs read to be consistent with
       // logic in handleHotUpdate()
       // for non vue files, e.g. md files in vitepress, we assume
-      // `hmrContext.read` is overwriten so handleHotUpdate() is dealing with
+      // `hmrContext.read` is overwritten so handleHotUpdate() is dealing with
       // post-transform code, so we populate the descriptor with post-transform
       // code here as well.
       filename.endsWith('.vue') ? undefined : code,
@@ -148,9 +148,20 @@ export async function transformMain(
       `typeof __VUE_HMR_RUNTIME__ !== 'undefined' && ` +
         `__VUE_HMR_RUNTIME__.createRecord(_sfc_main.__hmrId, _sfc_main)`,
     )
+    output.push(
+      `import.meta.hot.on('file-changed', ({ file }) => {`,
+      `  __VUE_HMR_RUNTIME__.CHANGED_FILE = file`,
+      `})`,
+    )
     // check if the template is the only thing that changed
     if (prevDescriptor && isOnlyTemplateChanged(prevDescriptor, descriptor)) {
-      output.push(`export const _rerender_only = true`)
+      // #7 only consider re-render if the HMR is triggered by the current component,
+      // otherwise reload. Due to vite will cache the transform result. If the HMR
+      // is triggered by other files that the current component relies on, a reload
+      // is required.
+      output.push(
+        `export const _rerender_only = __VUE_HMR_RUNTIME__.CHANGED_FILE === ${JSON.stringify(normalizePath(filename))}`,
+      )
     }
     output.push(
       `import.meta.hot.accept(mod => {`,